cleanup(batch19): CfSoundMan, CCur, CfGimmickObject, CfGimmickItem
- R1-R8: Full cleanup, headers created, no regressions

+++ src/kyoshin/cf/CfGimmickItem.ctx.c
@@ -1322,8 +1322,8 @@ public:
     static void func_8003AA8C(u32 val);
     static void* getFP(const char* pName) { return ::getFP(pName); }
     static const char* getBdatStringColumnValue(void* pData, const char* pColumnName, int index);
-    static u16 func_8003B1EC(void* pData);
-    static u16 func_8003B41C(void* pData);
+    static u32 func_8003B1EC(void* pData);
+    static u32 func_8003B41C(void* pData);
 };
 
 #ifdef __cplusplus
